split selection from selector_state
